x86/mm: Maintain the correct percpu mappings on context switch
Ensure the pagetables we are switching to have the correct percpu mappings in
them.  The _PGC_inuse_pgtable check ensures that the pagetables we edit aren't
in use elsewhere.

One complication however is context switching between two vcpus which both
require shadowing.  See the code comment for details.

Another complication is requiring a second percpu fixmap slot.  It limits
NR_CPUS to 254, but will be removed later.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>