x86/smp: Switch to using the percpu IDT mappings
The loading of IDTR is moved out of load_system_tables() and into
early_switch_to_idle().

One complication for the BSP is that IST references still need to remain
uninitalised until reinit_bsp_stack().  Therefore, early_switch_to_idle() is
extended to take a bsp boolean.

For VT-x guests, HOST_IDTR_BASE needs altering, so a #VMExit doesn't change
the mappings we use.  As this is now a compile-time constant, it moves from
vmx_set_host_env() to construct_vmcs() to avoid rewriting it on every context
switch.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>