x86/mapcache: Reimplement map_domain_page() from scratch
There are two reasons:
 1) To stop using the per-domain range for the mapcache
 2) To make map_domain_page() safe to use during context switches

The new implementation is entirely percpu and rather more simple.  See the
comment at the top of domain_page.c for a description of the algorithm.

A side effect of the new implementation is that we can get rid of struct
mapcache_{vcpu,domain} entirely, and mapcache_override_current().

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>

TODO: Consider whether to try and lazily unmap, utilising other TLB flush
scenarios rather than forcing an invlpg on each unmap.