x86/fixmap: Drop percpu fixmap range
The percpu fixmap range was introduced to allow opencoding of
map_domain_page() in the middle of a context switch.

The new implementation of map_domain_page() is safe to use in a context
switch, so drop the percpu fixmap infrastructure.

This removes the temporary build-time restriction on NR_CPUS.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>