x86/smp: Allocate a percpu linear range for the IDT
This change also introduces _alter_percpu_mappings(), a helper for creating
and modifying percpu mappings.  The code will be extended with extra actions
in later patches.

The existing IDT heap allocation and idt_tables[] array are kept, although the
allocation logic is simplified as an IDT is strictly one single page.

idt_table[], used by CPU0, now needs to be strictly page aligned, so is moved
into .bss.page_aligned.

Nothing writes to the IDT via its percpu mappings, so the opportunity is taken
to make the mapping read-only.  This provides extra defence-in-depth, as an
attacker can't use the pointer obtained from SIDT to modify the active IDT as
part of a privilege escalation attempt.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>