x86/smp: Use the percpu GDT/LDT mappings
This is unfortunately quite invasive, because of the impact on the context
switch path.

PV vcpus gain an array of ldt and gdt ptes (replacing gdt_frames[]), which map
the frames loaded by HYPERCALL_set_gdt, or faulted in for the LDT.  Each
present PTE here which isn't a read-only mapping of zero_page holds a type
reference.

When context switching to a vcpu which needs a full GDT or LDT, the ptes are
copied in from the above arrays, while if context switching away from vcpu
which needs a full GDT or LDT, the ptes are reset back to default values.  As
a side effect, the GDT/LDT create/destroy functions need to adjust the percpu
mappings if they are running in current context.

Overall, the GDT and LDT base addresses are always the same, and this depend
on the context switch logic happening before write_ptbase(), so the TLB flush
removes any stale mappings.  While altering load_LDT()'s behaviour to cope,
introduce lazy loading to avoid executing lldt in the common case.

The arch_{get,set}_info_guest() functions need adjusting to cope with the fact
that they will now find references to zero_page in the ptes, which need
skipping.

Loading of GDTR now happens once at boot, in early_switch_to_idle().  As the
base address is now constant and always mapped, we can remove lgdt() calls
from the context switch path and EFI Runtime Services path.  Finally,
HOST_GDTR_BASE in the VMCS needs to be adjusted, and moves into
construct_vmcs().

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
v2:
 * Fix pv_destroy_ldt() to release the correct references when there were
   outstanding LDT frames at domain destruction.