x86: Drop the PERDOMAIN mappings
With the mapcache, xlat and GDT/LDT moved over to the PERCPU mappings, there
are no remaining users of the PERDOMAIN mappings.  Drop the whole PERDOMAIN
infrastructure, and remove the PERDOMAIN slot in the virtual address layout.

Slide each of the subsequent slots back by one, and extend the directmap back
to its original size.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>