x86: Always reload the LDT on vcpu context switch
... and always zero the LDT for HVM contexts.  This causes erroneous execution
which manages to reference the LDT fail with a straight #GP fault, rather than
possibly finding a stale loaded LDT and wandering the #PF handler.

Future changes will cause the loading of LDT to be lazy, at which point
load_LDT() will be a nop for all cases other than context switching to/from a
PV vcpu with an LDT loaded.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>