x86/boot: Defer TSS/IST setup until later during boot on the BSP
TSS and IST setings are only required for safety when running userspace code.
Until we start executing dom0, the boot path is perfectly capable of handling
exceptions and interrupts without a loaded TSS.

Deferring the TSS setup is necessary to facilitiate moving the BSP onto a
percpu stack, which in turn requires that during boot, there are no IST
references in the IDT.

Correct TSS and IST settings are re-set up in reinit_bsp_stack(), just before
we complete initialisation.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>