x86/smp: Switch to using the percpu stacks
This is very easy for the APs.  __high_start() is modified to switch stacks
before entering C.  The BSP however is more complicated, and needs to stay on
cpu0_stack[] until setup is complete.

The end of __start_xen() is modified to copy the top-of-stack data to the
percpu stack immediately before jumping there.  The VMCS Host and SYSENTER
stacks are suitably adjusted, and become construction-time constant.

The stack_start and stack_base[] array are removed completely, as well as the
memguard_guard_stack() infrastructure.  The STACK_ORDER xenheap allocations
are no longer needed, and higher CPUs on large machines are finally
numa-local.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>