x86/misc: Move some IPI parameters off the stack
With percpu stacks, it will not be safe to pass stack pointers.  The logic in
machine_restart(), time_calibration() and set_mtrr() is singleton, so switch
to using static variables.

The set_mtrr_data is protected under the mtrr_mutex, which requires
mtrr_ap_init() and mtrr_aps_sync_end() to hold the mutex around calls to
set_mtrr().

time_calibration() runs exclusively out of a timer on cpu0 so is safe, while
machine_restart() doesn't have any concurrency to be worried about.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>