x86/smp: Use the percpu TSS mapping
Construction of the TSS is the final action remaining in load_system_tables(),
and is lifted to early_switch_to_idle().  As a single global TSS is in use,
the per_cpu init_tss variable is dropped.

The setting of HOST_TR_BASE is now a constant, so moves to construct_vmcs().
This means that vmx_set_host_env() and arch_vmx_struct.hostenv_migrated can be
dropped as well.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>