x86/amd: Introduce CPUID/MSR definitions for per-vcpu SSBD support
At the time of writing, the spec is available from:

  https://developer.amd.com/wp-content/resources/124441_AMD64_SpeculativeStoreBypassDisable_Whitepaper_final.pdf

Future hardware (Zen v2) is expect to have hardware MSR_SPEC_CTRL support,
including SPEC_CTRL.SSBD, and with the expectation that this will be directly
passed through to guests for performance.

On currently released hardware, the only mechanism available is the legacy
LS_CFG option, and this is very expensive to use.  Furthermore, emulating
MSR_SPEC_CTRL via interception is prohibitively expensive, as certain OSes use
the write-discard flexibility to simplify their entry/exit logic.

As an alternative, MSR_VIRT_SPEC_CTRL is specified as an architectural control
(with semantics equivilent to MSR_SPEC_CTRL) which is provided by the
hypervisor.  This abstracts away the model-specific details of the LS_CFG
mechanism, which allows migration safety to be retained.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
CC: Jan Beulich <JBeulich@suse.com>
CC: Wei Liu <wei.liu2@citrix.com>
CC: Roger Pau Monné <roger.pau@citrix.com>
CC: Brian Woods <brian.woods@amd.com>