x86/p2m: Don't overwrite p2m_altp2m_lazy_copy()'s callers p2m pointer
The final parameter to p2m_altp2m_lazy_copy() appears to be unnecessary, and
results in very hard-to-follow code.  Have the sole caller set its local p2m
pointer appropriately, and drop the parameter.

With that done, a level of indirection of ap2m can be dropped inside
p2m_altp2m_lazy_copy().  While changing the API, switch it from bool_t to
bool, and drop printing of the altp2m's virtual address, because it is of no
use even when debugging.

No (intended) overall change in behaviour.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Razvan Cojocaru <rcojocaru@bitdefender.com>
Reviewed-by: Jan Beulich <jbeulich@suse.com>
--
CC: Wei Liu <wei.liu2@citrix.com>
CC: Roger Pau Monné <roger.pau@citrix.com>
CC: Tamas K Lengyel <tamas@tklengyel.com>
CC: George Dunlap <george.dunlap@eu.citrix.com>