xen/mm: Switch mfn_to_virt()/virt_to_mfn() to using mfn_t
Seemingly, a majority of users either override the helpers anyway, or have an
mfn_t in their hands.

Update the API, and adjust all users to match.  In some places, use the
unsigned long variant in places where we are interacting with an external
struct and there is no chance of them switching to mfn_t.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Jan Beulich <jbeulich@suse.com>
Reviewed-by: Julien Grall <julien.grall@arm.com>
---
CC: Jan Beulich <JBeulich@suse.com>
CC: Wei Liu <wei.liu2@citrix.com>
CC: Roger Pau Monné <roger.pau@citrix.com>
CC: Stefano Stabellini <sstabellini@kernel.org>
CC: Julien Grall <julien.grall@arm.com>