x86/amd: Virtualise MSR_VIRT_SPEC_CTRL for guests
The semantics of MSR_VIRT_SPEC_CTRL are that unknown bits are write-discard
and read as zero.  Only VIRT_SPEC_CTRL.SSBD is defined at the moment.

To facilitate making this per-guest, the legacy SSBD state needs context
switching between vcpus.  amd_ctxt_switch_legacy_ssbd() is updated to take the
vcpus setting into account.  Furthermore, the guests chosen value needs
preserving across migrate.

This marks a subtle change in how `ssbd=` behaves.  If Xen wishes SSBD to be
asserted, it remains set in hardware all the time.  In the default case of Xen
wishing SSBD not to be asserted, the value set in hardware is the guests
choice.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
CC: Jan Beulich <JBeulich@suse.com>
CC: Wei Liu <wei.liu2@citrix.com>
CC: Roger Pau Monné <roger.pau@citrix.com>
CC: Brian Woods <brian.woods@amd.com>