xen/vcpu: Sanitise VCPUOP_initialise call hierachy
This code is especially tangled.  VCPUOP_initialise calls into
arch_initialise_vcpu() which calls back into default_initialise_vcpu() which
is common code.

This path is actually dead code on ARM, because VCPUOP_initialise is filtered
out by do_arm_vcpu_op().

The only valid way to start a secondary CPU on ARM is via the PSCI interface.
The same could in principle be said about INIT-SIPI-SIPI for x86 HVM, if HVM
guests hadn't already interited a paravirt way of starting CPUs.

Either way, it is quite likely that no future architectures implemented in Xen
are going to want to use a PV interface, as some standardised (v)CPU bringup
mechanism will already exist.

Arrange the code in do_vcpu_op() to allow arch_initialise_vcpu() to be
optional.  Opt in for x86, and opt out for ARM.

Deleting ARM's arch_initialise_vcpu() allows for default_initialise_vcpu() to
be folded into its (now) sole x86 caller, which reduces the compiled code
volume in all builds.

No functional change.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
CC: Jan Beulich <JBeulich@suse.com>
CC: Wei Liu <wl@xen.org>
CC: Roger Pau Monné <roger.pau@citrix.com>
CC: Stefano Stabellini <sstabellini@kernel.org>
CC: Julien Grall <julien@xen.org>
CC: Volodymyr Babchuk <Volodymyr_Babchuk@epam.com>
CC: Juergen Gross <jgross@suse.com>

This is XSA-296 followup, hence why it is only posted now.  Seeing as we are
fairly early in 4.13, I'd request that it be considered, but it won't be the
end of the world if it gets delayed for 4.14.